requirements.txt
disnake==2.9.3
humanize
pytz
millify
git+https://github.com/raianah/easy-pil
python-dotenv
disnake-ext-menus
aiomysql
# apscheduler